drm/i915: Include crtc contributed bits in plane_state->ctl
Currently we precompute plane_state->ctl with just the plane_state
contributed bits, and the crtc contributed bits are added in ad-hoc
fashion in all the places that write the plane control register.
Clean this up by also including the crtc bits in the precomputed
value.

Technically the only place that needs the _ctl() vs. _ctl_crtc()
split after this is i9xx_plane_disable_arm() (since we have to
write the crtc bits even when disabling the plane), but I've opted
to stick to the split on all platforms anyway. I think having clear
areas of responsibility for each function is nice.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>