drm/i915: Include plane ctl/surf registers in the plane update_arm() tracepoint
The plane fault tracepoints report the plane control and
surface register values. In order to correlate those against the
plane update tracepoints it might be helpful to also include that
information in the plane update tracepoints as well.

The one caveat here is that the precomputed ctl/surf values
that we include the tracepoint do not include the async flip
bit(s) (as those are handled in a more dynamic fashion), whereas
the ones read from the hardware in the plane fault tracepoint
include everything.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>